# spec/api.openapi.yaml
openapi: 3.1.0
info:
  title: Hello App VPods delivery API
  version: 1.0.0
  description: Application contracts published by Knox (backend).
paths:
...
  /api/orders:
    get:
      operationId: listOrders
      summary: List brownie catalog and persisted orders
      tags: [orders]
      responses:
        '200':
          description: Catalog and order queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersPayloadResponse'
    post:
      operationId: createOrder
      summary: Place a brownie batch order
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error
components:
  schemas:
...
    CreateOrderRequest:
      type: object
      required: [productName, quantity]
      properties:
        productName:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    OrderResponse:
      type: object
      required: [id, productName, quantity, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CatalogItemResponse:
      type: object
      required: [sku, name, unitPriceCents]
      properties:
        sku:
          type: string
        name:
          type: string
        unitPriceCents:
          type: integer
    OrdersPayloadResponse:
      type: object
      required: [catalog, orders]
      properties:
        catalog:
          type: array
          items:
            $ref: '#/components/schemas/CatalogItemResponse'
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderResponse'